Create new winstates branch and start to modify stuff
So at this time, the [api] branch has a broken makeMove function. I
broke it when I tried to add the feature to check for winstates
inside of the makeMove function and it became a monster. Even after I
opened an issue, It would only make each selection an 'x' and it would
say 'TIE' on each move. My curly brackets were out of control so I made
a new branch called winstates to abstract that winstate code out
and make a function to check for a winstate. Then I want to call that
function inside of the makeMove function. I am even making a new file
called winstates.js that I am requiring in the gameState array,
and exporting the checkWin function to the makeMove function. [api] was
supposed to be for the api but I ended up getting off track and
creating some good game logic, only to break it. Now the branch (which
is pushed up to my remote repo, what a mess) called api actually
has game logic that I expanded on and then eventually broke. Hoping to
fix this logic in [winstates] branch and merge winstates to [api] branch