Fix events.js to have partial functionality
Ok so I successfully reverted back to this state:
With the winstates removed from the code, I can click on different
tiles. The tiles will show x's and o's, alternating. If I try to
take a space that already has an x or o, then I get an error message
that goes away after I make a correct move. It doesn't know if the
game ends in any way.