Refactor code that checks for win states
So I went back through and reworked it so I didn't have to have 2
separate blocks to check if x and o have won. I wrote it so that it
would just check if each row was filled with the same player
regardless of which player it is. I just rewrote it in winstates.js to
mess around with syntax. The next step will be to copy and paste it
into events.js to try to call it in the onMove function.