Change empty gameState array back to 9 empty strings
I had a completely empty array in gameState array. My winstate
conditionals were looking for empty strings in each index of the array.
I just need to be careful about my conditionals and what they are
actually checking for. Empty strings are not the same as nothing!

Now my game won't say win or tie on every turn. It will correctly
log 'no winner yet, keep playing' to the console if there isn't a
winner. It will successfully call out a winner if a winstate
conditional is met. Next might be either tie state, or not let
moves to me made after gameOver.