Add functionality to stop the game after it ends
I added a gameOver variable and set it to false to start the game.
After each winstate, I reassign the gameOver variable to true and
this triggers a return at the beginning of the onMove function. If
gameOver is true, it returns out of the onMove function with
console.logs and messages reminding the user that the game is over
and they can click the New Game button to start a new game.

The New Game button has no functionality yet. Also, the message clears
the winner so it might be good to have a message to remind the user
who won and that they can start a new game by pressing the new game
button. I could do this by declaring a variable in each winstate
and assigning it the message function. I could then call this variable
in the message gameOver return at the beginning of the onMove
function.