Fix bug
I accidentally fixed a bug on my master branch. This worked out though.
When It was o's turn, and you clicked on an already taken space, it
would switch back to x's turn. So it would have x go twice.
I added a return after the error message and this seemed to fix the
problem.

I need to go over my branches and make sure they are up to date.