Add functionality to recognize a tie
I added another condition to the winstate function to look for each
index to not be empty and if that was true while no other winstates
were true, it would end the game in a tie. I log this to the console
and I updated the message to tell the user that there is a tie.